RPA.jl now works. bubble diagram calculation is much faster!
