--- docs/src/lib/diagram.md
@@ -1,5 +1,0 @@
-# Simple Diagrams
-
-```@autodocs
-Modules = [QuantumStatistics.Diagram]
-```
