+++ examples/RPA.jl
@@ -20,7 +20,7 @@ Note that this dynamic contribution ``dW_0'' diverges at small q. For this reaso
 function dWRPA(qgrid, τgrid, kF, β, spin, mass, e)
     @assert all(qgrid .!= 0.0)
     # instantaneous interaction (Coulomb interaction)
-    vq = [4π * e^2 / (q^2 + 0.001) for q in qgrid]
+    vq = [4π * e^2 / (q^2) for q in qgrid]
     EF = kF^2 / (2mass)
     dlr = DLR.DLRGrid(:corr, 10EF, β, 1e-10) # effective interaction is a correlation function of the form <O(τ)O(0)>
     Nq, Nτ = length(qgrid), length(τgrid)
@@ -28,18 +28,10 @@ function dWRPA(qgrid, τgrid, kF, β, spin, mass, e)
     dW0norm = similar(Π)
     for (ni, n) in enumerate(dlr.n)
         for (qi, q) in enumerate(qgrid)
-            iωn = 2π * n / β * im
-            Π[qi, ni] = Diagram.bubble(3, q, iωn, kF, β, mass)[1] .* spin
+            Π[qi, ni] = TwoPoint.LindhardΩnFiniteTemperature(3, q, n, kF, β, mass, 2)[1]
         end
         dW0norm[:, ni] = @. vq * Π[:, ni] / (1 - vq * Π[:, ni])
         println("ω_n=2π/β*$(n), Π(q=0, n=0)=$(Π[1, ni])")
-    end
-    println(Π[10, :])
-    for (qi, q) in enumerate(qgrid)
-        println("$qi  $q    $(dW0norm[qi, 1])")
-    end
-    for (ni, n) in enumerate(dlr.n)
-        println("$ni  $n   $(dW0norm[10, ni])")
     end
     dW0norm = DLR.matfreq2tau(:corr, dW0norm, dlr, τgrid, axis=2)
     return vq, real.(dW0norm)
@@ -54,11 +46,10 @@ if abspath(PROGRAM_FILE) == @__FILE__
 
     qgrid = Grid.boseK(kF, 3kF, 0.2kF, 32) 
     τgrid = Grid.tau(β, EF, 128)
-    println("qGrid: ", qgrid.grid)
-    println("τGrid: ", τgrid.grid)
+    # println("qGrid: ", qgrid.grid)
+    # println("τGrid: ", τgrid.grid)
 
     vq, dW0norm = dWRPA(qgrid.grid, τgrid.grid, kF, β, spin, m, e)
-    println(dW0norm[1, :])
     display(plot(qgrid.grid ./ kF, dW0norm[:, 1]))
     sleep(100)
     # plot(qgrid.grid ./ kF, dW0norm[:, 1])
